+++ blakserv/makefile.linux
@@ -49,6 +49,7 @@ OBJS =  \
 	$(OUTDIR)/commcli.obj \
 	$(OUTDIR)/string.obj \
 	$(OUTDIR)/async.obj \
+   $(OUTDIR)/async_linux.obj \
 	$(OUTDIR)/loadgame.obj \
 	$(OUTDIR)/game.obj \
 	$(OUTDIR)/term.obj \
